requirements.txt
albumentations==2.0.6
fastapi==0.115.12
matplotlib==3.10.3
numpy==2.2.5
opencv_contrib_python==4.10.0.84
opencv_python_headless==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
psycopg2_binary==2.9.10
pydantic==1.10.11
PyJWT==2.7.0
PyJWT==2.9.0
PyJWT==2.10.1
python_bcrypt==0.3.2
scikit_learn==1.6.1
streamlit==1.43.2
timm==1.0.15
torch==2.4.1
torch==2.4.1+cu118
torchvision==0.19.1
torchvision==0.19.1+cu118
tqdm==4.66.4
Werkzeug==2.2.3
